feat: 支持设置配置类型
