+++ nonebot_plugin_datastore/__init__.py
@@ -1,3 +1,7 @@
+from nonebot import require
+
+require("nonebot_plugin_localstore")
+
 from .db import create_session as create_session
 from .db import get_session as get_session
 from .plugin import PluginData as PluginData
