+++ nonebot_plugin_datastore/config.py
@@ -1,12 +1,12 @@
 """ 配置 """
 from pathlib import Path
-from typing import Any, Dict
+from typing import Any, Dict, Type
 
-from nonebot import get_driver, require
-from pydantic import BaseModel, Extra, root_validator
-
-require("nonebot_plugin_localstore")
+from nonebot import _resolve_dot_notation, get_driver
 from nonebot_plugin_localstore import get_cache_dir, get_config_dir, get_data_dir
+from pydantic import BaseModel, Extra, PrivateAttr, root_validator
+
+from .providers import ConfigProvider
 
 
 class Config(BaseModel, extra=Extra.ignore):
@@ -21,6 +21,13 @@ class Config(BaseModel, extra=Extra.ignore):
     datastore_enable_database: bool = True
     datastore_database_echo: bool = False
     datastore_engine_options: Dict[str, Any] = {}
+    datastore_config_provider: str
+
+    @property
+    def config_provider(self) -> Type[ConfigProvider]:
+        return self._config_provider
+
+    _config_provider: Type[ConfigProvider] = PrivateAttr()
 
     @root_validator(pre=True, allow_reuse=True)
     def set_defaults(cls, values: Dict):
@@ -48,6 +55,16 @@ class Config(BaseModel, extra=Extra.ignore):
             values[
                 "datastore_database_url"
             ] = f"sqlite+aiosqlite:///{values['datastore_data_dir'] / 'data.db'}"
+
+        # 设置默认配置提供器
+        # 默认使用 JSON
+        if not values.get("datastore_config_provider"):
+            values["datastore_config_provider"] = "~json"
+        cls._config_provider = _resolve_dot_notation(
+            values["datastore_config_provider"],
+            default_attr="Config",
+            default_prefix="nonebot_plugin_datastore.providers.",
+        )
         return values
